2008-06-10  Vladimir Makarov  <vmakarov@redhat.com>
	* ira-int.h (allocno_copy):  Change comments.
	
	* ira-conflicts.c (build_conflict_bit_table): Use cast in printf
	for %ld.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/ira@136637 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/ira-conflicts.c
@@ -113,8 +113,8 @@ build_conflict_bit_table (void)
     fprintf
       (ira_dump_file,
        "+++Allocating %ld bytes for conflict table (uncompressed size %ld)\n",
-       allocated_words_num * sizeof (INT_TYPE),
-       allocno_set_words * allocnos_num * sizeof (INT_TYPE));
+       (long) allocated_words_num * sizeof (INT_TYPE),
+       (long) allocno_set_words * allocnos_num * sizeof (INT_TYPE));
   for (i = 0; i < max_point; i++)
     {
       for (r = start_point_ranges[i]; r != NULL; r = r->start_next)
